// src/types/responses/offers/offers-list.ts
import {CityDto} from '../city-dto.ts';
import {Location} from '../../location.ts';

export type OffersList = OffersListItem[];

export type OffersListItem = {
  id: string;
  title: string;
  type: string;
  price: number;
  city: CityDto;
  location: Location;
  isFavorite: boolean;
  isPremium: boolean;
  rating: number;
  previewImage: string;
}